moved code component tests, fixing test issues
