+++ components/activity/code/test/d2l-activity-code-editor.test.js
@@ -1,0 +1,186 @@
+import '../d2l-activity-code-editor.js';
+import '../custom/d2l-activity-code-editor-learning-path.js';
+import { addToMock, mockLink } from '../../../../test/data/fetchMock.js';
+import { assert, html } from '@open-wc/testing';
+import { createComponentAndWait, fireEventAndWait } from '../../../../test/test-util.js';
+import { learningPathExisting, learningPathMissingAction, learningPathNew } from '../../../../test/data/learningPath.js';
+import { clearStore } from '@brightspace-hmc/foundation-engine/state/HypermediaState.js';
+import { runConstructor } from '@brightspace-ui/core/tools/constructor-test-helper.js';
+import sinon from 'sinon/pkg/sinon-esm.js';
+
+const inputText = 'd2l-input-text';
+
+async function _createCodeEditor(path) {
+	return await createComponentAndWait(html`<d2l-activity-code-editor href="${path}" token="test-token"></d2l-activity-code-editor>`);
+}
+
+async function _createCodeEditorLearningPath(path) {
+	return await createComponentAndWait(html`<d2l-activity-code-editor-learning-path href="${path}" token="test-token"></d2l-activity-code-editor-learning-path>`);
+}
+
+async function updateCode(element, updatedText) {
+	const inputArea = element.shadowRoot.querySelector(inputText);
+	inputArea.value = updatedText;
+
+	await fireEventAndWait(inputArea, 'input', element);
+}
+
+describe('d2l-activity-code-editor', async() => {
+
+	before(async() => {
+		mockLink.reset();
+		await addToMock('/learning-path/new', learningPathNew, _createCodeEditorLearningPath);
+		await addToMock('/learning-path/existing', learningPathExisting, _createCodeEditorLearningPath);
+		await addToMock('/learning-path/missing-action', learningPathMissingAction, _createCodeEditorLearningPath);
+	});
+	after(() => {
+		mockLink.reset();
+	});
+
+	describe('constructor', () => {
+
+		it('should construct', () => {
+			runConstructor('d2l-activity-code-editor');
+		});
+	});
+
+	describe('Component', () => {
+
+		beforeEach(() => {
+			clearStore();
+		});
+
+		afterEach(() => {
+			mockLink.resetHistory();
+		});
+
+		it('should initialize using defined path and expected values', async() => {
+			const element = await _createCodeEditor('/learning-path/new');
+
+			assert.equal(element.code, learningPathNew.properties.code);
+		});
+
+		describe('path:/learning-path/existing', () => {
+			let element;
+			beforeEach(async() => {
+				clearStore();
+				element = await _createCodeEditor('/learning-path/existing');
+				assert.equal(element.code, learningPathExisting.properties.code, 'code should match response');
+			});
+
+			it('code should be set when one is present', () => {
+				assert.equal(element.shadowRoot.querySelector(inputText).value,
+					learningPathExisting.properties.code, 'input value does not match');
+
+				assert.equal(element.code, learningPathExisting.properties.code, 'code property should match');
+
+			});
+
+			it('updating should commit state', async() => {
+				const spy = sinon.spy(element.updateCode);
+				await updateCode(element, 'new code');
+
+				assert.equal(element.code, 'new code', 'code was updated to match');
+				assert.isTrue(spy.commit.called, 'onInputCode should be called when input event is triggered');
+			});
+
+			it('extra whitespace is removed', async() => {
+				const spy = sinon.spy(element.updateCode);
+				await updateCode(element, '   new  code     ');
+
+				assert.equal(element.code, 'new  code', 'code was updated to match');
+				assert.isTrue(spy.commit.called, 'commit should be called when input event is triggered');
+			});
+		});
+		describe('path:/learning-path/missing-action', () => {
+			let element;
+			beforeEach(async() => {
+				clearStore();
+				element = await _createCodeEditor('learning-path/missing-action');
+				assert.equal(element.code, learningPathMissingAction.properties.code, 'code should match response');
+			});
+			it('code not updated when action missing', async() => {
+				const spy = sinon.spy(element.updateCode);
+				await updateCode(element, 'new code');
+
+				assert.equal(element.code, learningPathMissingAction.properties.code, 'code was unaltered');
+				assert.isFalse(spy.commit.called, 'commit should not be called on update failure');
+			});
+		});
+	});
+
+	describe('d2l-activity-code-editor-learning-path', () => {
+
+		describe('constructor', () => {
+
+			it('should construct', () => {
+				runConstructor('d2l-activity-code-editor-learning-path');
+			});
+		});
+
+		describe('Component', () => {
+
+			beforeEach(() => {
+				clearStore();
+			});
+
+			afterEach(() => {
+				mockLink.resetHistory();
+			});
+
+			it('should initialize using defined path and expected values', async() => {
+				const element = await _createCodeEditorLearningPath('/learning-path/new');
+
+				assert.equal(element.code, learningPathNew.properties.code);
+			});
+
+			describe('path:/learning-path/existing', () => {
+				let element;
+				beforeEach(async() => {
+					clearStore();
+					element = await _createCodeEditorLearningPath('/learning-path/existing');
+					assert.equal(element.code, learningPathExisting.properties.code, 'code should match response');
+				});
+
+				it('code should be set when one is present', () => {
+					assert.equal(element.shadowRoot.querySelector(inputText).value,
+						learningPathExisting.properties.code, 'input value does not match');
+
+					assert.equal(element.code, learningPathExisting.properties.code, 'code property should match');
+
+				});
+
+				it('updating should commit state', async() => {
+					const spy = sinon.spy(element.updateCode);
+					await updateCode(element, 'new code');
+
+					assert.equal(element.code, 'new code', 'code was updated to match');
+					assert.isTrue(spy.commit.called, 'onInputCode should be called when input event is triggered');
+				});
+
+				it('extra whitespace is removed', async() => {
+					const spy = sinon.spy(element.updateCode);
+					await updateCode(element, '   new  code     ');
+
+					assert.equal(element.code, 'new  code', 'code was updated to match');
+					assert.isTrue(spy.commit.called, 'commit should be called when input event is triggered');
+				});
+			});
+			describe('path:/learning-path/missing-action', () => {
+				let element;
+				beforeEach(async() => {
+					clearStore();
+					element = await _createCodeEditorLearningPath('learning-path/missing-action');
+					assert.equal(element.code, learningPathMissingAction.properties.code, 'code should match response');
+				});
+				it('code not updated when action missing', async() => {
+					const spy = sinon.spy(element.updateCode);
+					await updateCode(element, 'new code');
+
+					assert.equal(element.code, learningPathMissingAction.properties.code, 'code was unaltered');
+					assert.isFalse(spy.commit.called, 'commit should not be called on update failure');
+				});
+			});
+		});
+	});
+});
